+++ opentrons-ai-client/src/components/molecules/InputPrompt/inputprompt.module.css
@@ -5,7 +5,7 @@
 
 .error_container {
   display: flex;
-  align-items: flex-start;
+  align-items: center;
   padding: var(--spacing-12) var(--spacing-16);
   border: 1px solid var(--red-30);
   border-radius: var(--border-radius-4);
@@ -15,14 +15,13 @@
 }
 
 .error_icon {
-  width: 1rem;
-  height: 1rem;
   flex-shrink: 0;
-  margin-top: 0.125rem;
+  color: var(--red-50);
 }
 
-.error_text {
+.error_text_wrapper {
   flex: 1;
+  min-width: 0;
 }
 
 .error_dismiss {
@@ -30,7 +29,7 @@
   flex-shrink: 0;
   align-items: center;
   justify-content: center;
-  padding: 0;
+  padding: var(--spacing-8) var(--spacing-16);
   border: none;
   background: none;
   cursor: pointer;
