+++ opentrons-ai-client/src/resources/utils/index.ts
@@ -2,4 +2,6 @@ export * from './buildChatHistory'
 export * from './buildRequestConfig'
 export * from './calcTextAreaHeight'
 export * from './createUserInput'
+export * from './isApiErrorResponse'
 export * from './labware'
+export * from './resolveErrorMessage'
